requirements.txt
python-telegram-bot[ext]==20.7
kubernetes==28.1.0
aiohttp==3.9.1
PyYAML==6.0.1